remove unused class, rename file
